% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{addClustersToGatingSet}
\alias{addClustersToGatingSet}
\title{addClustersToGatingSet}
\usage{
addClustersToGatingSet(table, gs, parent_gate, fsom_file = NULL)
}
\arguments{
\item{table}{A data.table or tibble containing expression data for all samples.
Must have a column named \code{.id} indicating sample ID, and a column
\code{Metacluster} with cluster labels.}

\item{gs}{The \code{GatingSet} to add gates to.}

\item{parent_gate}{\code{character} indicating which population is the parent of
the clusters.}

\item{fsom_file}{Optional, a character giving a path to an .rds file for a FlowSOM
object. Default is \code{NULL}, which checks for a filename in \code{attr(table, "fsom_filename")}

This function should be used after a satisfactory clustering has been obtained
using the human-in-the-loop approach outlined by the vignette. The resulting
table should contain expression data for each sample, with each cell assigned
a cluster label. These labels will then be used to construct a gate for each
cluster.}
}
\description{
Uses cluster labels to add corresponding gates to the GatingSet
}
