% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{addMetadataToFlowSOM}
\alias{addMetadataToFlowSOM}
\title{addMetadataToFlowSOM}
\usage{
addMetadataToFlowSOM(fsom, aggregate, parent = NULL)
}
\arguments{
\item{fsom}{A FlowSOM object to annotate.}

\item{aggregate}{The aggregate .fcs file used to create the given FlowSOM object.}

\item{parent}{A filepath to a parent FlowSOM object's .rds file, if the given
FlowSOM object is a reclustering}
}
\value{
A FlowSOM object with added metadata.
}
\description{
addMetadataToFlowSOM
}
\keyword{internal}
