% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{addMetadataToGatingSet}
\alias{addMetadataToGatingSet}
\title{addMetadataToGatingSet}
\usage{
addMetadataToGatingSet(gs, sample_df)
}
\arguments{
\item{gs}{A \code{GatingSet}}

\item{sample_df}{A \code{data.frame}, where each row corresponds to a sample and
each column corresponds to metadata about the samples, such as experimental group,
or the filename of a corresponding control sample. Must contain a column for
filename, called \code{File.Name}.}
}
\description{
addMetadataToGatingSet
}
