% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{addMetadataToTable}
\alias{addMetadataToTable}
\title{addMetadataToTable}
\usage{
addMetadataToTable(table, sample_dt, join_col)
}
\arguments{
\item{table}{Table containing expression data for all samples}

\item{sample_dt}{Table containing sample info}

\item{join_col}{Column on which to join the two tables described above}
}
\description{
addMetadataToTable
}
