% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{calculateSE}
\alias{calculateSE}
\title{calculateSE}
\usage{
calculateSE(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Standard error.
}
\description{
Helper function for calculating standard error.
}
\keyword{internal}
