% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{clusterControls.data.frame}
\alias{clusterControls.data.frame}
\title{clusterControls.data.frame}
\usage{
\method{clusterControls}{data.frame}(
  input,
  fsom,
  subsetted_fsom = NULL,
  subsetted_meta = NULL,
  num_cells = nrow(fsom$data),
  save_fsom = NULL
)
}
\arguments{
\item{input}{A data.table, as generated by \code{\link[=getTableFromFCS]{getTableFromFCS()}} or \code{\link[=doPreprocessing]{doPreprocessing()}}.}

\item{fsom}{The original FlowSOM object, which has not been reclustered.}

\item{subsetted_fsom}{If the FlowSOM object of interest is a reclustering,
the reclustered FlowSOM object.}

\item{subsetted_meta}{If the FlowSOM object of interest is a reclustering,
the name of the metacluster that was used.}

\item{num_cells}{The number of cells to use in the resulting clustering.}

\item{save_fsom}{Optional, an .rds filename to be used to name the resulting
FlowSOM object. Default is \code{NULL}, which does not save the object.}
}
\value{
A data table with metacluster and cluster assignments for each cell.
}
\description{
Cluster FMO and/or isotype controls to facilitate calculation of delta MFIs.
}
