% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{clusterControls.FlowSOM}
\alias{clusterControls.FlowSOM}
\title{clusterControls.FlowSOM}
\usage{
\method{clusterControls}{FlowSOM}(
  fsom,
  sample_df,
  ctrl_col,
  dir_prepr_ctrl,
  marker,
  subsetted_meta = NULL,
  dir_clustr_ctrl = NULL,
  parent_ctrl_fsom = NULL
)
}
\arguments{
\item{fsom}{A FlowSOM object, to be used for differential expression analysis.}

\item{sample_df}{A data frame of sample info, as generated by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{ctrl_col}{The name of the column containing control filenames.}

\item{dir_prepr_ctrl}{The directory containing preprocessed control files.}

\item{marker}{The marker that this control corresponds to.}

\item{subsetted_meta}{A character vector specifying the metaclusters that were
used for metaclusters, if \code{fsom} is a reclustering. Default is \code{NULL}.}

\item{dir_clustr_ctrl}{A filepath to the directory containing clustered control
files, if \code{fsom} is a reclustering. Default is \code{NULL}.}

\item{parent_ctrl_fsom}{The parent control FlowSOM whose cells are to be
reclustered, if \code{fsom} is a reclustering. Default is \code{NULL}.}
}
\value{
A FlowSOM object.
}
\description{
Prepare isotype or FMO controls for calculating delta MFI.
}
