% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{clusterControls}
\alias{clusterControls}
\title{clusterControls}
\usage{
clusterControls(control_gs, primary_gs, population)
}
\arguments{
\item{control_gs}{A GatingSet containing control samples}

\item{primary_gs}{A GatingSet containing primary data}

\item{population}{The population that is to be clustered on}
}
\description{
clusterControls
}
