% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{clusterControlsTable}
\alias{clusterControlsTable}
\title{clusterControlsTable}
\usage{
clusterControlsTable(input, fsom, subsetted_meta = NULL, save_fsom = NULL)
}
\arguments{
\item{input}{A data.table.}

\item{fsom}{A FlowSOM object to map \code{input} to.}

\item{subsetted_meta}{If your FlowSOM object of interest is a reclustering,
the name of the metacluster that was used.}

\item{save_fsom}{Optional, an .rds filename to be used to name the resulting
FlowSOM object. Default is \code{NULL}, which does not save the object.}
}
\value{
A data table with metacluster and cluster assignments for each cell.
}
\description{
clusterControlsTable
}
