% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{clusterSubsetWithPCA}
\alias{clusterSubsetWithPCA}
\title{clusterSubsetWithPCA}
\usage{
clusterSubsetWithPCA(
  input,
  pca_obj,
  num_components,
  num_clus,
  xdim = 10,
  ydim = 10,
  seed = NULL,
  convert_to_channels = TRUE
)
}
\arguments{
\item{input}{An aggregated .fcs file, as generated by \code{\link[=createFilteredAggregate]{createFilteredAggregate()}},
or a data frame.}

\item{pca_obj}{A \code{\link[stats]{prcomp}} object as generated by \code{\link[=doPCA]{doPCA()}}}

\item{num_components}{The number of principal components to use for reclustering,
as determined by examining results of \code{\link[=plotPCAScree]{plotPCAScree()}}.}

\item{num_clus}{The number of metaclusters you would like.}

\item{xdim}{x-dimension of the SOM. Default is 10.}

\item{ydim}{y-dimension of the SOM. Default is 10.}

\item{seed}{A seed for reproducible results. Default is \code{NULL}.}

\item{convert_to_channels}{Logical, should the returned FlowSOM object's data
be defined in terms of channels. Default is \code{TRUE}.}
}
\value{
A FlowSOM object.
}
\description{
Recluster a subset of flow data using principal components of a filtered
aggregate .fcs file.
}
\details{
When reclustering a subset of flow data, such as a cell population that has
been identified as monocytes, not every marker you used for the initial clustering
is still relevant. In such a case, you would no longer be interested in using,
for instance, a B cell marker. One solution is to leave it up to the
researcher to determine what markers are appropriate for the reclustering,
but this approach is very user-dependent, therefore lacking reproducibility.

This function solves this problem by using PCA for reclustering. It
takes in an aggregate .fcs file, as generated by \code{createFilteredAggregate()},
and clusters the cells within it using the specified number of principal
components. If \code{convert_to_channels} is \code{TRUE} (default), it then
takes the cluster and metacluster assignments found from this clustering to
create a FlowSOM object defined in terms of channels, rather than
principal components, so that it may still be used with both the functions
in flowFun and in FlowSOM.
}
