% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{createFilteredAggregate.data.frame}
\alias{createFilteredAggregate.data.frame}
\title{createFilteredAggregate.data.frame}
\usage{
\method{createFilteredAggregate}{data.frame}(
  input,
  num_cells,
  clusters = NULL,
  metaclusters = NULL,
  dir_save = NULL
)
}
\description{
Filter data frame to create a new aggregate file.
}
\keyword{internal}
