% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{createFilteredAggregate.FlowSOM}
\alias{createFilteredAggregate.FlowSOM}
\title{createFilteredAggregate.FlowSOM}
\usage{
\method{createFilteredAggregate}{FlowSOM}(
  input,
  num_cells,
  clusters = NULL,
  metaclusters = NULL,
  agg_name = "filtered_aggregate.fcs",
  dir_save = NULL
)
}
\description{
Filter FlowSOM object to create a new aggregate file.
}
\keyword{internal}
