% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{createFilteredAggregate}
\alias{createFilteredAggregate}
\title{createFilteredAggregate}
\usage{
createFilteredAggregate(
  input,
  num_cells,
  clusters = NULL,
  metaclusters = NULL,
  agg_name = "aggregate.fcs",
  dir_save = NULL
)
}
\arguments{
\item{input}{A FlowSOM object as generated by \code{\link[FlowSOM:FlowSOM]{FlowSOM()}}, or
a data frame.}

\item{num_cells}{The total number of cells to include in the aggregate file.}

\item{clusters}{A numeric vector listing the clusters of interest.}

\item{metaclusters}{A character vector listing the metaclusters of interest.}

\item{agg_name}{An .fcs filename. Default is "filtered_aggregate.fcs".}

\item{dir_save}{If a directory path is given, then the filtered .fcs
files used to create the aggregate file will be saved there. If \code{NULL}
(default), then they will be discarded.}
}
\value{
An aggregated \code{\link[flowCore]{flowFrame}} or data table, where only cells
belonging to the given metaclusters and/or clusters are included.
}
\description{
Filter preprocessed data to create a new aggregate file.
}
\details{
It is only necessary to define one of the arguments \code{clusters} and
\code{metaclusters}, but the user may use both if they wish.

This function takes in either a data frame or FlowSOM object, and subsets
each sample to contain only the cells included in the given metaclusters
and/or clusters. If desired, new .fcs files for each of these subsetted samples
are then written to a new directory, whose name is specified by the user with
the \code{dir_save} parameter. Note that it is only helpful to save the
files if you intend to perform additional reclustering on them, otherwise
they may be safely deleted.
}
