% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{doAggregatePCA}
\alias{doAggregatePCA}
\title{doAggregatePCA}
\usage{
doAggregatePCA(aggregate, markers_to_cluster)
}
\arguments{
\item{aggregate}{An aggregate .fcs file as generated by \code{createFilteredAggregate}.}

\item{markers_to_cluster}{A character vector of the markers you would like to
use for reclustering.}
}
\value{
A list with class \code{prcomp}.
}
\description{
Performs PCA on an aggregate .fcs file's data matrix.
}
