% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{doDAAnalysis}
\alias{doDAAnalysis}
\title{doDAAnalysis}
\usage{
doDAAnalysis(
  design,
  counts,
  contrasts,
  sample_df,
  norm_method = c("TMM", "TMMwsp", "RLE", "upperquartile", "none")
)
}
\arguments{
\item{design}{A design matrix as generated by \code{\link[=makeDesignMatrix]{makeDesignMatrix()}}.}

\item{counts}{A count matrix as generated by \code{\link[=makeCountMatrix]{makeCountMatrix()}}.}

\item{contrasts}{A contrasts matrix as generated by \code{\link[=makeContrastsMatrix]{makeContrastsMatrix()}}.}

\item{sample_df}{A data frame, generated either manually or by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{norm_method}{The normalization method to be used and passed to
\code{\link[edgeR:calcNormFactors]{calcNormFactors()}}, one of \verb{"TMM", "TMMwsp", "RLE", "upperquartile", or "none"}. Default is \code{"TMM"}. See documentation
of \code{calcNormFactors()} for more details.}
}
\value{
A list of tables extracted from a
\code{\link[edgeR:`TopTags-class`]{TopTags}} object. Each element
corresponds to a column in \code{contrasts}.
}
\description{
Do differential abundance analysis on the data.
}
\details{
This function performs tests for differential abundances between groups of
interest defined by a contrasts matrix. Models are fit and compared via
likelihood ratio tests using functions from edgeR.
}
\seealso{
\code{\link[edgeR:calcNormFactors]{calcNormFactors()}},
\code{\link[edgeR:DGEList]{DGEList()}}, \code{\link[edgeR:estimateDisp]{estimateDisp()}},
\code{\link[edgeR:glmFit]{glmFit()}}, \code{\link[edgeR:glmLRT]{glmLRT()}},
\code{\link[edgeR:topTags]{topTags()}}.
}
