% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{doDEAnalysis.data.frame}
\alias{doDEAnalysis.data.frame}
\title{doDEAnalysis.data.frame}
\usage{
\method{doDEAnalysis}{data.frame}(
  input,
  sample_df,
  design,
  contrasts,
  counts,
  markers_of_interest,
  meta_names = NULL,
  ctrl_list = NULL
)
}
\description{
doDEAnalysis.data.frame
}
\keyword{internal}
