% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{doDEAnalysis.FlowSOM}
\alias{doDEAnalysis.FlowSOM}
\title{doDEAnalysis.FlowSOM}
\usage{
\method{doDEAnalysis}{FlowSOM}(
  input,
  sample_df,
  design,
  contrasts,
  counts,
  markers_of_interest,
  meta_names = NULL,
  prepr_transform = NULL,
  controls_df = NULL,
  ctrl_fsom_names = NULL,
  subsetted_meta = NULL
)
}
\arguments{
\item{prepr_transform}{A filepath to the .rds file for the logicle
transformation applied during preprocessing. Default is
\code{file.path("RDS", "logicle_transformation.rds")}.}

\item{controls_df}{Optional, a data frame as generated by \code{\link[=prepareControlInfo]{prepareControlInfo()}},
containing information about FMO or isotype controls to be used to calculate
delta MFIs.}

\item{ctrl_fsom_names}{Optional, a character string defining names of FlowSOM
.rds files previously generated from \code{\link[=clusterControls]{clusterControls()}}.}

\item{subsetted_meta}{If \code{fsom} is a reclustering, and \code{controls_df}
is not \code{NULL}, a string or list of strings defining the metacluster(s)
\code{fsom} is a reclustering of.}
}
\description{
Perform differential expression analysis on markers of interest.
}
