% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{doDEAnalysis}
\alias{doDEAnalysis}
\title{doDEAnalysis}
\usage{
doDEAnalysis(
  input,
  marker_cols,
  design,
  contrasts,
  subpopulations = NULL,
  inverse = FALSE
)
}
\arguments{
\item{input}{A \code{GatingSet} or \code{data.table}}

\item{marker_cols}{A \code{character} vector specifying which markers/channels
should be tested for differential expression. Each element should be a column
in the data's expression matrix.}

\item{design}{A design matrix from \code{\link[=makeDesignMatrix]{makeDesignMatrix()}}}

\item{contrasts}{A contrasts matrix from \code{\link[=makeContrastsMatrix]{makeContrastsMatrix()}}}

\item{subpopulations}{Only valid if \code{input} is a \code{GatingSet}. A \code{character} vector
specifying which populations in the \code{GatingSet} to include in tests.}

\item{inverse}{A \code{boolean}, whether or not data should be transformed back to
a linear scale before calculating MFIs and performing tests. Only valid if \code{input}
is a \code{GatingSet} with a transformation and its inverse attached.}
}
\value{
An \code{MArrayLM} object resulting from \code{\link[limma:ebayes]{limma::eBayes()}}
}
\description{
doDEAnalysis
}
