% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{doDEAnalysisTable}
\alias{doDEAnalysisTable}
\title{doDEAnalysis}
\usage{
doDEAnalysisTable(
  input,
  sample_df,
  design,
  contrasts,
  counts,
  markers_of_interest,
  meta_names = NULL,
  ctrl_list = NULL
)
}
\arguments{
\item{input}{A data.table.}

\item{sample_df}{A data frame of sample info, as generated by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{design}{A design matrix, as generated by \code{\link[=makeDesignMatrix]{makeDesignMatrix()}}.}

\item{contrasts}{A contrasts matrix, as generated by \code{\link[=makeContrastsMatrix]{makeContrastsMatrix()}}.}

\item{counts}{A count matrix, as generated by \code{\link[=makeContrastsMatrix]{makeContrastsMatrix()}}.}

\item{markers_of_interest}{A character vector of markers to test for differential
expression.}

\item{meta_names}{A character vector of metacluster names to test. Default is
all metaclusters, \code{levels(fsom$metaclustering)}.}
}
\value{
A list of data frames.
}
\description{
doDEAnalysis
}
