% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{doPCA.flowFrame}
\alias{doPCA.flowFrame}
\title{doPCA.flowFrame}
\usage{
\method{doPCA}{flowFrame}(input, cols_to_use)
}
\description{
Performs PCA on an aggregate .fcs file's data matrix.
}
\keyword{internal}
