% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{doPCA}
\alias{doPCA}
\title{doPCA}
\usage{
doPCA(input, cols_to_use)
}
\arguments{
\item{input}{A data frame.}

\item{cols_to_use}{Columns of the input to use for PCA.}
}
\value{
A prcomp object.
}
\description{
Performs PCA on a data.frame or a flowFrame's data matrix.
}
