% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{doPreprocessing}
\alias{doPreprocessing}
\title{doPreprocessing}
\usage{
doPreprocessing(
  input,
  ld_channel,
  compensation = NULL,
  transformation = NULL,
  debris_gate = NULL,
  live_gate = NULL,
  nmad = 4,
  pctg_live = 0.6,
  pctg_qc = 0.8,
  pdf_name = "preprocessing_results.pdf",
  save_fcs = FALSE
)
}
\arguments{
\item{input}{A directory, list of filenames, or data.table.}

\item{ld_channel}{Name of the channel corresponding to the marker for live/dead
cells. This should appear the same as it does in the .fcs files.}

\item{compensation}{Optional. A compensation matrix, or a file path to one.}

\item{transformation}{The transformation to apply to the data. If \code{NULL}
(default), a log-icle transformation is applied to the data.}

\item{debris_gate}{A gate to gate out debris, defined as a
\code{\link[flowCore:filter]{filter()}}. If \code{NULL} (default), a
gate is automatically selected with \code{\link[flowDensity:flowDensity]{flowDensity()}}.}

\item{live_gate}{A gate to gate out dead cells. If \code{NULL} (default), a
gate is automatically selected.}

\item{nmad}{Parameter to determine strictness of doublet removal. See
\code{\link[PeacoQC:RemoveDoublets]{PeacoQC::RemoveDoublets()}} for details.}

\item{pctg_live}{Minimum percentage of live cells a sample should have.}

\item{pctg_qc}{Minimum percentage of cells a sample should have after removing
low-quality events during quality control.}

\item{pdf_name}{Name of the PDF containing diagnostic plots for preprocessing
results. Default is \code{"preprocessing_results.pdf"}.}

\item{save_fcs}{Boolean, should the preprocessed data be saved as .fcs files.
Default is \code{FALSE}.}
}
\value{
A data.table containing preprocessed data.
}
\description{
Preprocess all files in the given directory.
!!! WIP edit gates and handling of low-quality samples, rename markers if necessary
}
\details{
The steps taken by this function, in order, are:
\enumerate{
\item Removing margin events (i.e. values that fall outside the range the
flow cytometer should be able to pick up).
\item Removing doublets.
\item Removing debris.
\item Compensating the data.
\item Performing a log-icle transformation on each channel.
\item Removing dead cells.
\item Quality control, to check and correct for occurrences like clogs, speed
changes, etc. and flag any files that may be too low-quality to include
in the analysis.
}

The script will output a few files to check for any mistakes. The first is
a PDF highlighting the cells that were removed in each .fcs file. The others
are plots for any files that were flagged during quality control, where each
channel is plotted against time and removed events are marked.
}
