% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{editGateManual}
\alias{editGateManual}
\title{editGateManual}
\usage{
editGateManual(gs, node, dims = NULL, ref_sample = 1, sample_ids = NULL)
}
\arguments{
\item{gs}{A \code{GatingSet} whose gates will be manually edited}

\item{node}{A \code{character} string specifying which node to redraw}

\item{dims}{A \code{character} vector indicating which channels to draw the new
gate on. By default, it is set to the channels used in the original gate. If it
is of length-1, a histogram will be displayed for the user to draw on. If
length-2, a scatterplot will be displayed instead.}

\item{ref_sample}{A \code{numeric} or \code{character}, specifying the index
or name of the sample within the \code{GatingSet} that you would like to
draw the new gate on. NOTE: if \code{sample_ids} is specified by the user,
\code{ref_sample} should be found within the vector \code{sample_ids}.
(e.g. if I have a GatingSet
of length 10, and set \code{sample_ids = 6:10}, then \code{ref_sample} should
be a number between 6 and 10.)}

\item{sample_ids}{Optional; if you would like to redraw this gate for only a subset
of your samples, a vector of sample indices or names.}
}
\value{
A \code{GatingSet} whose gates have been redrawn according to user input.
}
\description{
A function enabling the user to redraw gates for all or a subset of a
\code{GatingSet}
}
