% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{editTableMetaclusters}
\alias{editTableMetaclusters}
\title{editTableMetaclusters}
\usage{
editTableMetaclusters(
  table,
  new_labels = NULL,
  cluster_assignments = NULL,
  level_order = NULL
)
}
\arguments{
\item{table}{A data table with columns for "Metacluster" and "Cluster".}

\item{new_labels}{Optional. A named vector, where names are the original metacluster names,
and values are the new labels.}

\item{cluster_assignments}{Optional. A named vector, where names are the cluster numbers,
and values are metacluster names.}

\item{level_order}{A character vector giving the desired order of the metacluster labels.}
}
\value{
A data.table with an added column \code{Meta_edited}.
}
\description{
Merge or relabel metaclusters, and reassign clusters.
}
\examples{
file <- system.file("extdata", "fsom_table_init.rds", package = "flowFun")
fsom_table <- readRDS(file)

fsom_table <- editTableMetaclusters(fsom_table,
                                    new_labels = c(
                                    "9" = "1",
                                    "2" = "1",
                                    "8" = "4",
                                    "23" = "15",
                                    "21" = "15",
                                    "20" = "15",
                                    "22" = "15",
                                    "7" = "6"),
                                    cluster_assignments = c("4" = "Undefined")
                                    )

print(fsom_table)
}
