% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{filterData.data.frame}
\alias{filterData.data.frame}
\title{filterData.data.frame}
\usage{
\method{filterData}{data.frame}(input, clusters = NULL, metaclusters = NULL)
}
\value{
A tidytable, where only cells belonging to the given clusters or
metaclusters are included.
}
\description{
filterData.data.frame
}
\keyword{internal}
