% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{filterData.flowFrame}
\alias{filterData.flowFrame}
\title{filterData.flowFrame}
\usage{
\method{filterData}{flowFrame}(input, clusters = NULL, metaclusters = NULL)
}
\value{
A flowFrame, where only cells belonging to the given clusters or
metaclusters are included.
}
\description{
filterData.flowFrame
}
\keyword{internal}
