% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{filterFCS}
\alias{filterFCS}
\title{filterFCS}
\usage{
filterFCS(ff, clusters = NULL, metaclusters = NULL)
}
\arguments{
\item{ff}{A flowFrame.}

\item{clusters}{A character vector specifying which clusters to keep. Default
is \code{NULL}.}

\item{metaclusters}{A character vector specifying which metaclusters to keep.
Default is \code{NULL}.}
}
\value{
A flowFrame, where only cells belonging to the given clusters or
metaclusters are included.
}
\description{
filterFCS
}
