% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{flagChannelNames}
\alias{flagChannelNames}
\title{flagChannelNames}
\usage{
flagChannelNames(input)
}
\arguments{
\item{input}{A flowSet or cytoset}
}
\value{
A list. For each sample, contains a matrix of channel/marker pairs
that were inconsistent, i.e. not found in all samples
}
\description{
Checks for typos/discrepancies between a flowSet or cytoset's channel names
between samples
}
