% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{flowSOMToTable}
\alias{flowSOMToTable}
\title{flowSOMToTable}
\usage{
flowSOMToTable(fsom)
}
\arguments{
\item{fsom}{}
}
\value{
A tidytable
}
\description{
Convert FlowSOM object to a tidytable
}
