% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_agnostic.R
\name{flowSOMWrapper}
\alias{flowSOMWrapper}
\title{flowSOMWrapper}
\usage{
flowSOMWrapper(
  table,
  markers_to_cluster,
  num_clus,
  seed = NULL,
  fsom_file = "flowsom_object.rds",
  ...
)
}
\arguments{
\item{table}{...}

\item{markers_to_cluster}{...}

\item{num_clus}{...}

\item{seed}{...}

\item{fsom_file}{...}

\item{...}{...}
}
\value{
A data.table annotated with cluster and metacluster assignments.
}
\description{
flowSOMWrapper
}
