% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gatingSetToTable}
\alias{gatingSetToTable}
\title{gatingSetToTable}
\usage{
gatingSetToTable(gs, population)
}
\arguments{
\item{gs}{Preprocessed GatingSet to convert}

\item{population}{A string giving the name of the gated population to use for
table creation}
}
\value{
A data.table, created from the given GatingSet
}
\description{
gatingSetToTable
}
