% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{generateGatingTable}
\alias{generateGatingTable}
\title{generateGatingTable}
\usage{
generateGatingTable(gs, collapse_data = TRUE, ld_stain = NULL)
}
\arguments{
\item{gs}{A \code{GatingSet}}

\item{collapse_data}{A \code{boolean}, indicating if samples should be collapsed
into one. Default is \code{TRUE}}

\item{ld_stain}{If the experiment contains a live/dead stain, a \code{character}
indicating which channel}
}
\value{
A data.table, to be used for creation of a \code{gatingTemplate}
}
\description{
generateGatingTable
}
\seealso{
\code{\link[openCyto:gatingTemplate-class]{openCyto::gatingTemplate()}}
}
