% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{getChannel}
\alias{getChannel}
\title{getChannel}
\usage{
getChannel(ff, marker)
}
\arguments{
\item{ff}{A flowFrame.}

\item{marker}{The marker of interest.}
}
\value{
A channel name.
}
\description{
Helper for getting a flowFrame marker's corresponding channel.
}
