% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getChannelMarkerPairs}
\alias{getChannelMarkerPairs}
\title{getChannelMarkerPairs}
\usage{
getChannelMarkerPairs(ff, save_res = TRUE)
}
\arguments{
\item{ff}{A flowFrame or filepath to a .fcs file}

\item{save_res}{Boolean; should the results be saved to the current working directory}
}
\value{
A two column table specifying channel/marker pairs
}
\description{
...
}
