% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getClusterIndicesBySample}
\alias{getClusterIndicesBySample}
\title{getClusterIndicesBySample}
\usage{
getClusterIndicesBySample(table)
}
\value{
A list of tibbles, where each element corresponds to a sample and
contains the indices of cells belonging to each cluster
}
\description{
getClusterIndicesBySample
}
