% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{getClusterMFIs}
\alias{getClusterMFIs}
\title{getClusterMFIs}
\usage{
getClusterMFIs(input, cols_to_use = NULL)
}
\arguments{
\item{input}{A data frame or table, with column \code{"Cluster"}.}

\item{cols_to_use}{A vector specifying which columns to calculate medians for.
Default is all columns used for clustering.}
}
\value{
A data frame, where rows are clusters and columns are channels.
}
\description{
getClusterMFIs
}
