% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{getDensity}
\alias{getDensity}
\title{getDensity}
\usage{
getDensity(x, y, ...)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{...}{Additional parameters to pass to \code{\link[MASS:kde2d]{MASS::kde2d()}}.}
}
\value{
A matrix of the estimated density.
}
\description{
Helper for plotting by density
}
\keyword{internal}
