% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getFullNames}
\alias{getFullNames}
\title{Function to match a character vector whose elements are either markers or
channels to a parent vector whose elements contain both markers and
the corresponding channel. i.e. match "CD3" to "CD3 <Alexa Fluor 700-A>".}
\usage{
getFullNames(substr, target_names, keep_names = FALSE)
}
\description{
Function to match a character vector whose elements are either markers or
channels to a parent vector whose elements contain both markers and
the corresponding channel. i.e. match "CD3" to "CD3 <Alexa Fluor 700-A>".
}
\keyword{internal}
