% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{getMarker}
\alias{getMarker}
\title{getMarker}
\usage{
getMarker(ff, channel)
}
\arguments{
\item{ff}{A flowFrame.}

\item{channel}{The channel of interest.}
}
\value{
A marker name.
}
\description{
Helper for getting a flowFrame channel's corresponding marker.
}
