% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{getMetaclusterMFIs}
\alias{getMetaclusterMFIs}
\title{getMetaclusterMFIs}
\usage{
getMetaclusterMFIs(input, cols_to_use = NULL)
}
\arguments{
\item{input}{A data frame or table, with column \code{"Metacluster"}.}

\item{cols_to_use}{A vector specifying which columns to calculate medians for.
Default is all columns used for clustering.}
}
\value{
A data frame, where rows are metaclusters and columns are channels.
}
\description{
getMetaclusterMFIs
}
