% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getSampleMetaclusterMFIs}
\alias{getSampleMetaclusterMFIs}
\title{getSampleMetaclusterMFIs}
\usage{
getSampleMetaclusterMFIs(
  input,
  col,
  populations = NULL,
  transformation = NULL,
  inverse = FALSE
)
}
\arguments{
\item{input}{A \code{data.table} or \code{GatingSet}}

\item{col}{The channel to find sample-metacluster MFIs for.}

\item{populations}{A \code{character} list of the names of metaclusters/populations to calculate MFIs for.}

\item{transformation}{A \code{\link[flowCore:transformList]{transformList}} specifying
the transformation applied to the data, if any. Required when \code{input} is a \code{data.table}
and \code{inverse = TRUE}}

\item{inverse}{Boolean, whether data should be back-transformed before calculating MFIs.
Only valid when \code{input} is a \code{GatingSet} containing a transformation.}
}
\value{
A table, where rows are samples and columns are metaclusters.
}
\description{
Get a table of MFIs, where each row is a sample and each column is a metacluster.
}
