% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{getSampleMFIs}
\alias{getSampleMFIs}
\title{getSampleMFIs}
\usage{
getSampleMFIs(input)
}
\arguments{
\item{input}{A FlowSOM object, flowFrame, or matrix of expression values.
Must have a \code{"File"} column in the data matrix.}
}
\value{
A data frame, where each row is a sample, and each column is a channel/marker.
}
\description{
A helper function to get sample MFIs.
add functionality for tidytable
}
