% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{getTableFromFCS}
\alias{getTableFromFCS}
\title{getTableFromFCS}
\usage{
getTableFromFCS(input, num_cells = NULL)
}
\arguments{
\item{input}{List of filenames with relative file paths, or a directory name.}

\item{num_cells}{Optional parameter for stratified sampling, the number of
cells to randomly sample from each file.}
}
\value{
A data.table containing .fcs file data.
}
\description{
Generate a data.table to use for analysis from existing .fcs files.
}
