% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{gs_makeDeltaMFIs}
\alias{gs_makeDeltaMFIs}
\title{gs_makeDeltaMFIs}
\usage{
gs_makeDeltaMFIs(gs, control_gs, subpopulations, cols, metadata_col = NULL)
}
\arguments{
\item{gs}{A GatingSet to find delta MFIs for}

\item{control_gs}{A GatingSet containing control samples}

\item{subpopulations}{The populations to perform the calculation for. Passed to
\code{gs_makeMFIMatrix()}}

\item{cols}{The markers to find delta MFIs for}

\item{metadata_col}{The column in \code{pData(gs)} to use to match control and primary
samples. If \code{NULL}, samples in both \code{GatingSets} will be assumed to be in the
correct order}
}
\description{
gs_makeDeltaMFIs
}
