% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{gs_makeMFIMatrix}
\alias{gs_makeMFIMatrix}
\title{gs_makeMFIMatrix}
\usage{
gs_makeMFIMatrix(gs, cols, subpopulations, inverse = FALSE)
}
\arguments{
\item{gs}{A \code{GatingSet}}

\item{cols}{A \code{character} vector of the columns in the expression matrix
to calculate MFIs for}

\item{subpopulations}{A \code{character} vector of subpopulations in \code{gs}}

\item{inverse}{\code{boolean}, whether or not the data should be transformed back
to its raw scale. If \code{TRUE}, a transformation with an associated inverse must be attached
to the \code{GatingSet}}
}
\value{
A data.frame of MFIs
}
\description{
gs_makeMFIMatrix
}
