% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{inverseTransformTable}
\alias{inverseTransformTable}
\title{NEED TO HANDLE MISSING PARAMETERS FROM TRANSFORMATION}
\usage{
inverseTransformTable(input, transformation)
}
\description{
NEED TO HANDLE MISSING PARAMETERS FROM TRANSFORMATION
}
\keyword{internal}
