% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{loadFlowSOMEmbedding}
\alias{loadFlowSOMEmbedding}
\title{loadFlowSOMEmbedding}
\usage{
loadFlowSOMEmbedding(file, newData)
}
\arguments{
\item{file}{A file path to the FlowSOM object whose clusters you would like \code{newData}
to map to}

\item{newData}{A flowFrame, flowSet, matrix with column names, or a list of
paths for files or directories containing the data to be clustered}
}
\value{
A FlowSOM object
}
\description{
Load saved FlowSOM and apply to new data
}
