% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{makeBoolean}
\alias{makeBoolean}
\title{makeBoolean}
\usage{
makeBoolean(input, indices, keep_indices = FALSE)
}
\value{
A boolean vector where \code{TRUE} means a cell belongs to the population
}
\description{
makeBoolean
}
