% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeCountMatrix.data.frame}
\alias{makeCountMatrix.data.frame}
\title{makeCountMatrix.data.frame}
\usage{
\method{makeCountMatrix}{data.frame}(input, meta_names = NULL, min_cells = 3, min_samples = NULL)
}
\description{
Generate matrix of sample/metacluster cell counts.
}
\keyword{internal}
