% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeCountMatrix.FlowSOM}
\alias{makeCountMatrix.FlowSOM}
\title{makeCountMatrix.FlowSOM}
\usage{
\method{makeCountMatrix}{FlowSOM}(
  input,
  sample_df,
  meta_names = NULL,
  min_cells = 3,
  min_samples = NULL
)
}
\arguments{
\item{sample_df}{A data frame containing sample information, generated by
\code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}
}
\description{
Generate matrix of sample/metacluster cell counts from FlowSOM object.
}
\keyword{internal}
