% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeDesignMatrix}
\alias{makeDesignMatrix}
\title{makeDesignMatrix}
\usage{
makeDesignMatrix(sample_df)
}
\arguments{
\item{sample_df}{A data frame, generated either manually or by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}
}
\value{
A matrix where each column is a group of interest, and each row is a
sample.
}
\description{
Generate a design matrix. (edit example)
}
\examples{
filepath <- system.file("extdata", "sample_information.csv", package = "flowFun")
samples <- prepareSampleInfo(filepath, "Sample.Name", "File.Name", comparisons)

design <- makeDesignMatrix(samples)
}
