% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeFactorDF}
\alias{makeFactorDF}
\title{makeFactorDF}
\usage{
makeFactorDF(sample_info, comparisons)
}
\arguments{
\item{sample_info}{A list of lists containing sample information}

\item{comparisons}{A named list of named lists, defining the groups to be
compared during analysis. See example for how this variable should be defined.}
}
\value{
A data frame of factors, with an added column for group.
}
\description{
!!! original function now mostly merged with prepareSampleInfo(), only really
necessary if sampleinfo is given as a list
}
