% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{plotAllSamples}
\alias{plotAllSamples}
\title{plotAllSamples}
\usage{
plotAllSamples(gs, xdim, ydim, subset, node)
}
\arguments{
\item{gs}{A \code{GatingSet}}

\item{xdim}{The channel to plot on the x-axis}

\item{ydim}{The channel to plot on the y-axis}

\item{subset}{The population whose cells should be displayed on the plot}

\item{node}{The gate to be visualized}
}
\value{
A \code{ggplot} drawn with \code{ggcyto}
}
\description{
Plot chosen gate for all samples
}
