% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{plotBeforeAfter}
\alias{plotBeforeAfter}
\title{plotBeforeAfter}
\usage{
plotBeforeAfter(ff1, ff2, channel1, channel2, ncells)
}
\arguments{
\item{ff1}{The "before" flowFrame.}

\item{ff2}{The "after" flowFrame.}

\item{channel1}{The channel to plot on the x-axis.}

\item{channel2}{The channel to plot on the y-axis.}

\item{ncells}{The number of cells to include in the plot.}
}
\value{
A plot where cells that were removed are highlighted in red, and cells
that were kept are highlighted in blue.
}
\description{
Plot events that were removed between steps.
}
