% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotClusterFileProportions}
\alias{plotClusterFileProportions}
\title{plotClusterFileProportions}
\usage{
plotClusterFileProportions(fsom, view = c("MST", "grid"), reg_expr = NULL)
}
\arguments{
\item{fsom}{A FlowSOM object.}

\item{view}{A string specifying which version of the plot you would like returned
either "MST" or "grid".}

\item{reg_expr}{Optional, a regular expression to rename filenames.}
}
\value{
An MST or grid plot.
}
\description{
Plots file distributions by cluster in an MST or grid.
}
\details{
The pie chart in the bottom left may serve as a point of reference,
as it shows what one would expect the proportions to be if
there was no difference between samples, and if the clustering
went well. Note that if a metacluster or cluster is made up
mostly of one sample's cells, this does not necessarily mean that the
clustering is flawed, and may instead indicate that the relevant sample has
unique biological characteristics.
}
