% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotClusterGroupProportions}
\alias{plotClusterGroupProportions}
\title{plotClusterGroupProportions}
\usage{
plotClusterGroupProportions(fsom, groups, view = c("MST", "grid"))
}
\arguments{
\item{fsom}{A FlowSOM object as generated by \code{\link[FlowSOM]{FlowSOM}}.}

\item{groups}{A list of lists, specifying the groups of interest and their file
names.}

\item{view}{A string specifying which version of the plot you would like returned
either "MST" or "grid".}
}
\value{
An MST or grid plot.
}
\description{
Plot group proportions by cluster in an MST or grid.
}
\details{
For each cluster, a pie chart displaying the proportions
of each given group within said cluster is created. They
may be visualized as either an MST or a grid. The pie chart in
the bottom left may serve as a point of reference, as it shows what we would
expect the proportions to be if there was no difference between groups.
}
