% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotClusterMFIs}
\alias{plotClusterMFIs}
\title{plotClusterMFIs}
\usage{
plotClusterMFIs(
  fsom,
  markers_of_interest = fsom$map$colsUsed,
  metaclusters = NULL,
  ...
)
}
\arguments{
\item{fsom}{The FlowSOM object whose data you would like to plot.}

\item{markers_of_interest}{A character vector specifying which markers you
would like to include in the plot. Default is fsom$map$colsUsed (the markers
used for clustering).}

\item{metaclusters}{A vector specifying which metaclusters whose clusters
you would like to include in the plot. If NULL (default), all clusters are included.}

\item{...}{Additional parameters to pass to \code{ComplexHeatmap::Heatmap}.}
}
\value{
A heatmap of MFIs made with ComplexHeatmap, where each column is a
marker of interest, and each row is a cluster.
}
\description{
Plot a heatmap of cluster MFIs.
}
\examples{
file <- system.file("extdata", "sample_aggregate.fcs", package = "flowFun")
fsom <- FlowSOM::FlowSOM(file,
                         colsToUse = c(10, 12:14, 16, 18:23, 25:32, 34),
                         nClus = 10,
                         seed = 42,
                         xdim = 6,
                         ydim = 6)

# Plot all clusters
plotClusterMFIs(fsom)

# Plot only clusters belonging to metaclusters 9 and 10
plotClusterMFIs(fsom, metaclusters = c(9, 10))
}
