% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{plotGroupMFIBars}
\alias{plotGroupMFIBars}
\title{plotGroupMFIBars}
\usage{
plotGroupMFIBars(
  input,
  col,
  sample_df,
  comparison,
  populations = NULL,
  transformation = NULL,
  inverse = FALSE,
  upper_lim = NULL
)
}
\arguments{
\item{input}{A matrix or data.frame where rows are samples, columns are clusters,
and values are MFIs for a channel}

\item{col}{Character, the name of the marker/channel to create the plot for.}

\item{sample_df}{A factor data frame as generated by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{comparison}{A list defining a comparison.}

\item{populations}{A vector of strings defining the metaclusters to include
in the plot. Default is all metaclusters.}

\item{transformation}{A \code{\link[flowCore:transformList]{transformList}} specifying
the transformation applied to the data, if any. Required when \code{input} is a \code{data.table}
and \code{inverse = TRUE}}

\item{inverse}{Boolean, whether the data should be back-transformed before
plotting.}

\item{upper_lim}{The y-axis upper limit.}
}
\value{
A bar plot drawn with \code{\link[ggplot2]{ggplot2}}.
}
\description{
Function to draw bar plot of MFIs or dMFIs by group.
}
\details{
The names and order of the groups defined in \code{comparison} should
appear in the same way you would like them to appear in the bar plot. Make
sure that the group names you supply are the same as they appear in the
\code{sample_df} object.
}
