% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{plotGroupUMAPs}
\alias{plotGroupUMAPs}
\title{plotGroupUMAPs}
\usage{
plotGroupUMAPs(
  fsom,
  sample_df,
  comparison,
  umap = NULL,
  color_by = "density",
  num_cells = 5000,
  seed = NULL
)
}
\arguments{
\item{fsom}{A FlowSOM object.}

\item{sample_df}{A factor data frame as generated by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{comparison}{A list specifying the groups of interest, in terms
of file names.}

\item{umap}{A UMAP plot as generated by \code{\link[=plotUMAP]{plotUMAP()}}. If \code{NULL}
(default), a new UMAP will be generated for the plots.}

\item{color_by}{A string specifying how you would like the UMAP plot to be colored.
Your options are "density" or a marker or channel of interest. Default is "density".}

\item{num_cells}{The number of cells you would like to be sampled for each
group. Default is 5000.}

\item{seed}{Optional, a seed for reproducibility.}
}
\value{
Plots faceted by group drawn with \code{\link[ggplot2]{ggplot2}}.
}
\description{
Plot colored UMAPs for each group of interest.
}
\details{
If any group has fewer than \code{num_cells}, then the number of cells to use
for plotting will be taken from the group with the lowest cell count.

!!! add check for total cells being greater than number used to make fsom
when umap = NULL
}
