% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotLabeled2DScatter.FlowSOM}
\alias{plotLabeled2DScatter.FlowSOM}
\title{Note that the output of this function changes slightly depending on the combination
of clusters and metaclusters the user is plotting. If the user plots metaclusters,
with or without additional clusters defined in the \code{clusters} argument,
then the output will be a single ggplot object, where each metacluster has a
unique color. If the user plots only clusters, such that the \code{metaclusters}
argument remains \code{NULL}, then a list of ggplot objects will be returned.
Each plot will correspond to a cluster, and color only the cells belonging to
said cluster. See examples for further explanation.}
\usage{
\method{plotLabeled2DScatter}{FlowSOM}(
  input,
  channelpair,
  clusters = NULL,
  metaclusters = NULL,
  label_type = c("cluster", "metacluster")
)
}
\description{
The (meta)cluster number labels are included on every plot.
}
\details{
#' file <- system.file("extdata", "aggregate.fcs", package = "flowFun")
fsom <- FlowSOM::FlowSOM(file,
colsToUse = c(10, 12:14, 16, 18:23, 25:32, 34),
nClus = 10,
seed = 42,
xdim = 6,
ydim = 6)
}
\section{Returns a single plot colored by metacluster}{
plotLabeled2DScatter(fsom,
channelpair = c("APC-Cy7-A", "BUV615-P-A"),
metaclusters = c(1, 2, 7),
label_type = "metacluster")
}

\section{Also returns a single plot colored by metacluster}{
plotLabeled2DScatter(fsom,
channelpair = c("APC-Cy7-A", "BUV805-A"),
clusters = c(25, 33),
metaclusters = c(1, 4),
label_type = "cluster")
}

\section{Returns a list of two plots, once for each cluster}{
p = plotLabeled2DScatter(fsom,
channelpair = c("APC-Cy7-A", "BUV805-A"),
clusters = c(4, 19),
label_type = "cluster")
print(p[\link{1}])
print(p[\link{2}])
}

