% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotLabeled2DScatter}
\alias{plotLabeled2DScatter}
\title{plotLabeled2DScatter}
\usage{
plotLabeled2DScatter(
  fsom,
  channelpair,
  clusters = NULL,
  metaclusters = NULL,
  label_type = c("cluster", "metacluster")
)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by the FlowSOM function.}

\item{channelpair}{A vector of two channel names.}

\item{clusters}{A vector specifying indices of clusters of interest.}

\item{metaclusters}{A vector specifying indices of metaclusters of interest.}

\item{label_type}{A string specifying whether you would like the cluster
centers to be labeled with the cluster or metacluster they belong to, either
"cluster" or "metacluster".}
}
\value{
A list of scatterplots or single scatterplot made with ggplot and
\code{\link[FlowSOM:Plot2DScatters]{FlowSOM::Plot2DScatters()}}.
}
\description{
Plot 2D scatterplots of FlowSOM clusters, with (meta)cluster number labels for
each cluster center.
}
\details{
It is only necessary to define one of the arguments \code{clusters} and
\code{metaclusters}, but the user may use both if they wish.

Note that the output of this function changes slightly depending on the combination
of clusters and metaclusters the user is plotting. If the user plots metaclusters,
with or without additional clusters defined in the \code{clusters} argument,
then the output will be a single ggplot object, where each metacluster has a
unique color. If the user plots only clusters, such that the \code{metaclusters}
argument remains \code{NULL}, then a list of ggplot objects will be returned.
Each plot will correspond to a cluster, and color only the cells belonging to
said cluster. See examples for further explanation.

The (meta)cluster number labels are included on every plot.
}
\examples{
file <- system.file("extdata", "sample_aggregate.fcs", package = "flowFun")
fsom <- FlowSOM::FlowSOM(file,
                         colsToUse = c(10, 12:14, 16, 18:23, 25:32, 34),
                         nClus = 10,
                         seed = 42,
                         xdim = 6,
                         ydim = 6)

# Returns a single plot colored by metacluster
plotLabeled2DScatter(fsom,
                     channelpair = c("APC-Cy7-A", "BUV615-P-A"),
                     metaclusters = c(1, 2, 7),
                     label_type = "metacluster")

# Also returns a single plot colored by metacluster
plotLabeled2DScatter(fsom,
                     channelpair = c("APC-Cy7-A", "BUV805-A"),
                     clusters = c(25, 33),
                     metaclusters = c(1, 4),
                     label_type = "cluster")

# Returns a list of two plots, once for each cluster
p = plotLabeled2DScatter(fsom,
                         channelpair = c("APC-Cy7-A", "BUV805-A"),
                         clusters = c(4, 19),
                         label_type = "cluster")
print(p[[1]])
print(p[[2]])
}
