% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotMarkerDensityByChannel}
\alias{plotMarkerDensityByChannel}
\title{plotMarkerDensityByChannel}
\usage{
plotMarkerDensityByChannel(gs, channel, population = "root", inverse = FALSE)
}
\arguments{
\item{gs}{A \code{GatingSet}}

\item{channel}{The channel to plot densities for}

\item{population}{A \code{character} indicating which subpopulation to use}

\item{inverse}{A boolean, whether or not the data should be inverse transformed
before plotting. \code{FALSE} by default}
}
\value{
A faceted plot with density plots of the chosen channel for each sample
}
\description{
plotMarkerDensityByChannel
}
