% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotMetaclusterMFIs.data.frame}
\alias{plotMetaclusterMFIs.data.frame}
\title{plotMetaclusterMFIs.data.frame}
\usage{
\method{plotMetaclusterMFIs}{data.frame}(input, cols_to_use = NULL, ...)
}
\description{
plotMetaclusterMFIs.data.frame
}
\keyword{internal}
