% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotMetaclusterMFIs.FlowSOM}
\alias{plotMetaclusterMFIs.FlowSOM}
\title{plotMetaclusterMFIs.FlowSOM}
\usage{
\method{plotMetaclusterMFIs}{FlowSOM}(input, cols_to_use = NULL, ...)
}
\description{
plotMetaclusterMFIs.FlowSOM
}
\keyword{internal}
