% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{plotPCAScree}
\alias{plotPCAScree}
\title{plotPCAScree}
\usage{
plotPCAScree(pca_obj)
}
\arguments{
\item{pca_obj}{A \code{\link[stats]{prcomp}} object as generated by \link{doAggregatePCA}.}
}
\value{
A scree plot generated with \code{\link[ggplot2]{ggplot2}}.
}
\description{
Plot a scree plot for an aggregate .fcs file's principal components.
}
\details{
This function produces a scree plot, which may be used to determine which principal
components should be used in the analysis. Typically, there is a distinct
"elbow" in the plot, where the amount of variance explained by each component
becomes drastically smaller. So, if this elbow appears at principal component
5, you would use components 1-5 to recluster your data. If there is no
obvious elbow, it is instead reasonable to use whichever number of components
explains roughly 80\% of the variance.
}
