% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotSampleProportions}
\alias{plotSampleProportions}
\title{plotSampleProportions}
\usage{
plotSampleProportions(count_mat)
}
\arguments{
\item{count_mat}{A data frame or table, where rows are clusters and
columns are samples.}
}
\value{
A stacked bar plot made with \code{\link[ggplot2]{ggplot2}}.
}
\description{
Plot cell type proportions by sample.
}
\details{
By default, the names used for each sample will be its file name.
}
\seealso{
\code{\link[=getExprMatDE]{getExprMatDE()}}
}
