% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotUMAP.data.frame}
\alias{plotUMAP.data.frame}
\title{plotUMAP.data.frame}
\usage{
\method{plotUMAP}{data.frame}(input, num_cells = 5000, labels = NULL, colors = NULL, seed = NULL)
}
\arguments{
\item{input}{A data table or FlowSOM object.}

\item{num_cells}{The number of cells to use to generate the UMAP. Default is 5000.}

\item{seed}{Optional, a seed for reproducibility.}
}
\value{
A UMAP plot drawn with ggplot2.
}
\description{
plotUMAP.data.frame
}
