% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotUMAPOld}
\alias{plotUMAPOld}
\title{plotUMAPOld}
\usage{
plotUMAPOld(fsom, num_cells, point_size = 1.5, seed = NULL)
}
\arguments{
\item{fsom}{The FlowSOM object whose data you would like to plot.}

\item{num_cells}{The total number of cells you would like to use for the
dimension reduction.}

\item{point_size}{Optional parameter to adjust the size of the plotted points.
Default is 1.5.}

\item{seed}{Optional parameter to set a seed for reproducibility.}
}
\value{
A UMAP plot made with ggplot2.
}
\description{
Plot a UMAP dimension reduction of FlowSOM data, colored by metacluster.
}
\examples{
file <- system.file("extdata", "sample_aggregate.fcs", package = "flowFun")
fsom <- FlowSOM::FlowSOM(file,
                         colsToUse = c(10, 12:14, 16, 18:23, 25:32, 34),
                         nClus = 10,
                         seed = 42,
                         xdim = 6,
                         ydim = 6)

plotUMAPOld(fsom, num_cells = 2500)

plotUMAPOld(fsom, num_cells = 2500, point_size = 3, seed = 42)
}
