% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{prepareCompensationMatrix}
\alias{prepareCompensationMatrix}
\title{prepareCompensationMatrix}
\usage{
prepareCompensationMatrix(matrix, gs)
}
\arguments{
\item{matrix}{The compensation matrix to prepare}

\item{gs}{The \code{GatingSet} whose parameter names the compensation matrix's
columns should be matched to}
}
\description{
Changes columns of compensation matrix to channels, markers, or both (pretty)
and checks order of columns
}
