% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{prepareControlInfo}
\alias{prepareControlInfo}
\title{prepareControlInfo}
\usage{
prepareControlInfo(
  markers,
  ctrl_cols,
  ctrl_prepr_dirs,
  ctrl_clustr_dirs = NULL,
  parent_ctrl_fsom = NULL
)
}
\arguments{
\item{markers}{...}

\item{ctrl_cols}{...}

\item{ctrl_prepr_dirs}{...}

\item{ctrl_clustr_dirs}{...}

\item{parent_ctrl_fsom}{...}
}
\value{
A data frame with FMO/Isotype control information.
}
\description{
use in clusterControls ?
}
