% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{prepareControls}
\alias{prepareControls}
\title{prepareControls}
\usage{
prepareControls(
  sample_df,
  control_col,
  marker,
  orig_fsom,
  fsom_name = NULL,
  agg_name = NULL,
  subsetted_meta = NULL
)
}
\description{
Prepare isotype or FMO controls for calculating delta MFI.
}
\details{
Additional details...
}
