% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{prepareGateParams}
\alias{prepareGateParams}
\title{prepareGateParams}
\usage{
prepareGateParams(gate)
}
\arguments{
\item{gate}{A gate defined by either the flowDensity or flowCore packages, or a list.}
}
\description{
prepareGateParams
}
\keyword{internal}
