% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{previewPreprocessing.GatingSet}
\alias{previewPreprocessing.GatingSet}
\title{previewPreprocessing.GatingSet}
\usage{
previewPreprocessing.GatingSet(
  input,
  sample_ind = 1,
  compensation = NULL,
  transformation = c("logicle", "arcsinh", "linear"),
  ld_channel = NULL,
  debris_args = list(),
  singlet_args = list(),
  live_args = list()
)
}
\arguments{
\item{input}{A \code{GatingSet} created with the \code{flowWorkspace} package}

\item{sample_ind}{A character string or integer specifying which sample to examine;
default is \code{1}}

\item{compensation}{A numeric matrix}

\item{transformation}{A character string specifying which transformation to apply,
or a \code{transformList} defining a custom one}

\item{ld_channel}{Optional; a string specifying the name of the channel used
to detect live/dead cells}

\item{debris_args}{A string defining any additional gating arguments to use
for the non-debris gate; see \code{\link[openCyto:gate_mindensity]{openCyto::gate_mindensity()}}}

\item{singlet_args}{A string defining any additional gating arguments to use
for the non-debris gate; see \code{\link[flowStats:gate_singlet]{flowStats::gate_singlet()}}}

\item{live_args}{A string defining any additional gating arguments to use
for the live cell gate; see \code{\link[openCyto:gate_mindensity]{openCyto::gate_mindensity()}}}
}
\value{
A plot
}
\description{
...
}
