% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{previewPreprocessing}
\alias{previewPreprocessing}
\title{previewPreprocessing}
\usage{
previewPreprocessing(
  input,
  ld_channel = NULL,
  compensation = NULL,
  transformation = NULL,
  transformation_type = c("logicle", "arcsinh", "other", "none"),
  debris_gate = NULL,
  live_gate = NULL,
  nmad = 4
)
}
\arguments{
\item{input}{A filepath or flowFrame.}

\item{ld_channel}{Name of the channel corresponding to the marker for live/dead
cells. This should appear the same as it does in the .fcs files.}

\item{compensation}{Optional. A compensation matrix, or a file path to one.}

\item{transformation}{The transformation to apply to the data. If \code{NULL}
(default), a log-icle transformation is applied to the data.}

\item{transformation_type}{The type of transformation to be applied to the data.
If \code{transformation} is \code{NULL}, the transformation parameters will be
selected automatically. See details for more.}

\item{debris_gate}{A gate to gate out debris, defined as a
\code{\link[flowCore:filter-class]{filter}}. If \code{NULL} (default), a
gate is automatically selected with \code{\link[flowDensity:flowDensity-methods]{flowDensity()}}.}

\item{live_gate}{A gate to gate out dead cells. If \code{NULL} (default), a
gate is automatically selected.}

\item{nmad}{Parameter to determine strictness of doublet removal. See
\code{\link[PeacoQC:RemoveDoublets]{PeacoQC::RemoveDoublets()}} for details.}
}
\description{
Preview preprocessing settings for a single file before applying it to all data.
}
