% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{removeFilepath}
\alias{removeFilepath}
\title{removeFilepath}
\usage{
removeFilepath(filenames)
}
\arguments{
\item{filenames}{A list of filenames.}
}
\value{
A list of filenames.
}
\description{
Helper to remove filepath from a filename.
}
