% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{revertPCAFlowSOM}
\alias{revertPCAFlowSOM}
\title{revertPCAFlowSOM}
\usage{
revertPCAFlowSOM(fsom, aggregate)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by \code{clusterSubsetWithPCA}.}

\item{aggregate}{The aggregate .fcs file used to generate the given FlowSOM
object.}
}
\value{
A FlowSOM object defined in terms of channels.
}
\description{
Reverts elements of a FlowSOM object generated on principal components back
to channels.
}
\details{
This function converts a FlowSOM object created from principal components,
as generated by \code{clusterSubsetWithPCA}, into an
equivalent FlowSOM object defined in terms of channels instead. The
cluster/metacluster assignments for each cell remains the same, but the
resulting FlowSOM object is able to be used with built-in FlowSOM functions
like \code{FlowSOM::PlotStars}, \code{FlowSOM::PlotPies}, etc.
Therefore other flowFun functions like \code{plotLabeled2DScatter} and
\code{searchByExpression} also work on the resulting object.
}
