% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{saveFlowSOMEmbedding}
\alias{saveFlowSOMEmbedding}
\title{saveFlowSOMEmbedding}
\usage{
saveFlowSOMEmbedding(fsom, file)
}
\arguments{
\item{fsom}{The FlowSOM object whose mapping you would like to save}

\item{file}{A filepath or filename to save the resulting object under}
}
\description{
Save essential parts of FlowSOM
}
