% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{searchByExpression}
\alias{searchByExpression}
\title{searchByExpression}
\usage{
searchByExpression(fsom, levels)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by the FlowSOM function.}

\item{levels}{A list specifying channels of interest and whether they are
high or low. Note that searching for intermediate expression is not supported.
See examples for the proper format of this variable.}
}
\value{
An MST plot, where nodes with the given expression levels are
highlighted and labeled as "True".
}
\description{
Search MST for nodes with certain marker expressions.
}
\examples{
file <- system.file("extdata", "sample_aggregate.fcs", package = "flowFun")
fsom <- FlowSOM::FlowSOM(file,
                         colsToUse = c(10, 12:14, 16, 18:23, 25:32, 34),
                         nClus = 15,
                         seed = 42)

searchByExpression(fsom,
                   levels = c("BUV661-A" = "high",
                              "APC-Cy7-A" = "low"))

searchByExpression(fsom,
                   levels = c("BUV805-A" = "low",
                              "APC-Cy7-A" = "high",
                              "Alexa Fluor 700-A" = "high"))
}
