% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dir_clustr}
\alias{set_dir_clustr}
\title{Edit dir_clustr}
\usage{
set_dir_clustr(new_dir_clustr)
}
\arguments{
\item{new_dir_clustr}{d A string defining the new directory for clustered .fcs
files, as generated by \code{FlowSOM::SaveClustersToFCS}.}
}
\description{
Edit dir_clustr
}
