% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dir_edger}
\alias{set_dir_edger}
\title{Edit dir_edger}
\usage{
set_dir_edger(new_dir_edger)
}
\arguments{
\item{new_dir_edger}{A string defining the directory to store results of
differential abundance analysis from edgeR.}
}
\description{
Edit dir_edger
}
