% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dir_limma}
\alias{set_dir_limma}
\title{Edit dir_limma}
\usage{
set_dir_limma(new_dir_limma)
}
\arguments{
\item{new_dir_limma}{A string defining the directory to store results of
differential expression analysis from limma.}
}
\description{
Edit dir_limma
}
