% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dir_prepr}
\alias{set_dir_prepr}
\title{Edit dir_prepr}
\usage{
set_dir_prepr(new_dir_prepr)
}
\arguments{
\item{new_dir_prepr}{A string defining the new directory for preprocessed .fcs
files.}
}
\description{
Edit dir_prepr
}
