% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dir_rds_edited}
\alias{set_dir_rds_edited}
\title{Edit dir_rds_edited}
\usage{
set_dir_rds_edited(new_dir_rds_edited)
}
\arguments{
\item{new_dir_rds_edited}{A string defining the new directory for edited
RDS files.}
}
\description{
Edit dir_rds_edited
}
