% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{set_dir_rds_unedited}
\alias{set_dir_rds_unedited}
\title{Edit dir_rds_unedited}
\usage{
set_dir_rds_unedited(new_dir_rds_unedited)
}
\arguments{
\item{new_dir_rds_unedited}{A string defining the new directory for unedited
RDS files.}
}
\description{
Edit dir_rds_unedited
}
