% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{setupDirectories}
\alias{setupDirectories}
\title{setupDirectories}
\usage{
setupDirectories(dirs_custom = list())
}
\arguments{
\item{dirs_custom}{A named list specifying any custom directory names.}
}
\description{
!!! edit !!! Initializes and creates all directories necessary for the pipeline.
}
\details{
By default, this function creates a directory structure as follows:
\itemize{
\item \code{getwd()}
\itemize{
\item Data
\itemize{
\item SetUp (\code{dir_setup})
\item Preprocessed SetUp (\code{dir_setup_prepr})
\item Raw (\code{dir_raw})
\item Preprocessed Files (\code{dir_prepr})
\item Aggregates (\code{dir_agg})
\item Clustered Files (\code{dir_clustr})
}
\item Info
\item RDS (\code{dir_rds})
\itemize{
\item Unedited (\code{dir_rds_unedited})
\item Edited (\code{dir_rds_edited})
}
\item Preprocessing Results
\itemize{
\item QC plots
}
\item Analysis Results
\itemize{
\item limma Analysis
\item edgeR Analysis
}
}
}
}
\examples{
setupDirectories(list(dir_prepr = file.path("Data", "Preprocessed Files Batch1")))
}
