% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{startProject}
\alias{startProject}
\title{startProject}
\usage{
startProject(dir_name = "Cytometry_Analysis")
}
\arguments{
\item{dir_name}{Absolute or relative filepath giving the name of the directory
that you would like your analysis to be saved in. By default, a folder named
"Cytometry_Analysis" is created in the current working directory.}
}
\description{
Set working directory to a previous analysis, or create a new one.
??? anything else to instantiate here ???
}
