% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{transformTable}
\alias{transformTable}
\title{NEED TO HANDLE MISSING PARAMETERS FROM TRANSFORMATION}
\usage{
transformTable(
  input,
  transformation = NULL,
  transform_type = c("logicle", "arcsinh", "other", "none"),
  find_inverse = FALSE
)
}
\description{
NEED TO HANDLE MISSING PARAMETERS FROM TRANSFORMATION
}
\keyword{internal}
