% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{addMetadataToFlowSOM}
\alias{addMetadataToFlowSOM}
\title{addMetadataToFlowSOM}
\usage{
addMetadataToFlowSOM(fsom, aggregate)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by \code{\link[FlowSOM]{FlowSOM}}.}

\item{aggregate}{The aggregate .fcs file used to create the given FlowSOM object.}
}
\value{
A FlowSOM object with added metadata.
}
\description{
addMetadataToFlowSOM
}
