% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{annotateMFIHeatmap}
\alias{annotateMFIHeatmap}
\title{annotateMFIHeatmap}
\usage{
annotateMFIHeatmap(original_fsom, new_fsom, heatmap = NULL)
}
\arguments{
\item{original_fsom}{The initial FlowSOM object.}

\item{new_fsom}{The final FlowSOM object, whose metaclusters have been merged
and named via \code{\link[FlowSOM:UpdateMetaclusters]{FlowSOM::UpdateMetaclusters()}}.}

\item{heatmap}{Optional, a heatmap as generated by \code{\link[=plotMetaclusterMFIs]{plotMetaclusterMFIs()}}. If
\code{NULL} (default), then a new heatmap will be generated from \code{original_fsom}
and annotated, otherwise the given heatmap will be annotated.}
}
\value{
An annotated heatmap generated by \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}.
}
\description{
Annotate an unmerged heatmap of MFIs with each metacluster's final assignment.
}
\details{
Additional details...
}
