% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{clusterControls}
\alias{clusterControls}
\title{clusterControls}
\usage{
clusterControls(fsom, subsetted_meta = NULL, ...)
}
\arguments{
\item{fsom}{A FlowSOM object, to be used for differential expression analysis.}

\item{subsetted_meta}{A character vector specifying the metaclusters that were
used for metaclusters, if \code{fsom} is a reclustering. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A FlowSOM object.
}
\description{
Prepare isotype or FMO controls for calculating delta MFI.
}
\details{
Additional details...
}
