% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{clusterSubset}
\alias{clusterSubset}
\title{clusterSubset}
\usage{
clusterSubset(aggregate, fsom_name, markers_to_cluster, num_clus)
}
\arguments{
\item{aggregate}{An aggregated .fcs file, as generated by \code{\link[=createFilteredAggregate]{createFilteredAggregate()}}.}

\item{fsom_name}{A string to name the resulting FlowSOM object.}

\item{markers_to_cluster}{A character vector of the markers you would like to
use for reclustering.}

\item{num_clus}{The number of metaclusters you would like.}
}
\value{
A FlowSOM object as generated by \code{\link[FlowSOM:FlowSOM]{FlowSOM()}}.
}
\description{
consider deleting this one !!!
}
\details{
This function may be used if you wish to select the markers to use for
clustering yourself, rather than use PCA. It takes in an aggregate .fcs file,
and clusters the cells within it using the specified markers of interest.

The resulting FlowSOM object will be saved to the directory defined in
\code{dir_rds_unedit()} with the name specified by \code{fsom_name}.
}
