% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{createFilteredAggregate}
\alias{createFilteredAggregate}
\title{createFilteredAggregate}
\usage{
createFilteredAggregate(
  fsom,
  num_cells,
  clusters = NULL,
  metaclusters = NULL,
  agg_name = "filtered_aggregate.fcs",
  dir_save = NULL
)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by the FlowSOM function.}

\item{num_cells}{The total number of cells to include in the aggregate file.}

\item{clusters}{A numeric vector listing the clusters of interest.}

\item{metaclusters}{A character vector listing the metaclusters of interest.}

\item{agg_name}{An .fcs filename. Default is "filtered_aggregate.fcs".}

\item{dir_save}{If a directory path is given, then the filtered .fcs
files used to create the aggregate file will be saved there. If \code{NULL}
(default), then they will be discarded.}
}
\value{
An aggregated flowFrame, where only cells belonging to the given
metaclusters and/or clusters are included.
}
\description{
Filter preprocessed .fcs files to create a new aggregate file
}
\details{
It is only necessary to define one of the arguments \code{clusters} and
\code{metaclusters}, but the user may use both if they wish.

This function takes in a group of clustered .fcs files, as generated
by \code{FlowSOM::SaveClustersToFCS}, and subsets them to create a new set of
.fcs files that contain only the cells included in the given metaclusters
and/or clusters. If desired, these new .fcs files are then written to a new
directory, whose name is specified by the user with the \code{dir_save}
parameter. Note that it is only necessary to save the files if you intend to perform
additional reclustering on them, otherwise they may be safely deleted. The
aggregate file is written to the directory saved in \code{dir_agg()} under the
name specified by \code{agg_name}.
}
