% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{doDAAnalysis}
\alias{doDAAnalysis}
\title{doDAAnalysis}
\usage{
doDAAnalysis(
  design,
  counts,
  contrasts,
  factors,
  norm_method = c("TMM", "TMMwsp", "RLE", "upperquartile", "none"),
  dir_tables = NULL
)
}
\arguments{
\item{design}{A design matrix as generated by \code{\link[=makeDesignMatrix]{makeDesignMatrix()}}.}

\item{counts}{A count matrix as generated by \code{\link[=makeCountMatrix]{makeCountMatrix()}}.}

\item{contrasts}{A contrasts matrix as generated by \code{\link[=makeContrastsMatrix]{makeContrastsMatrix()}}.}

\item{factors}{A data frame, generated either manually or by \code{\link[=makeFactorDF]{makeFactorDF()}}.}

\item{norm_method}{The normalization method to be used and passed to
\code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}, if any.}

\item{dir_tables}{The name of a subdirectory to save the resulting tables in.
Default is \code{NULL}, in which case the files will be saved in \code{dir_edger}.}
}
\description{
Do differential abundance analysis on the data.
}
\details{
This function performs tests for differential abundances between groups of
interest defined by a contrasts matrix. Models are fit and compared via
likelihood ratio tests using functions from edgeR. Results are saved as a
.csv file for each comparison in the directory defined by \code{dir_edger()} at
runtime. If running an edgeR analysis on multiple FlowSOM objects, or running the
script multiple times for any reason, the user may wish to create
sub-directories within \code{dir_edger} to stay organized. To do this,
simply set the \code{dir_tables} parameter to a preferred sub-directory name.
}
\seealso{
\code{\link[edgeR:calcNormFactors]{edgeR::calcNormFactors()}}, \code{\link[edgeR:DGEList]{edgeR::DGEList()}}, \code{\link[edgeR:estimateDisp]{edgeR::estimateDisp()}},
\code{\link[edgeR:glmfit]{edgeR::glmFit()}}, \code{\link[edgeR:glmLRT]{edgeR::glmLRT()}}, \code{\link[edgeR:topTags]{edgeR::topTags()}}.
}
