% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{doDEAnalysis}
\alias{doDEAnalysis}
\title{doDEAnalysis}
\usage{
doDEAnalysis(
  fsom,
  sample_df,
  design,
  contrasts,
  counts,
  markers_of_interest,
  meta_names = NULL,
  prepr_transform = NULL,
  controls_df = NULL,
  ctrl_fsom_names = NULL,
  subsetted_meta = NULL
)
}
\arguments{
\item{fsom}{A FlowSOM object, as generated by \code{\link[FlowSOM]{FlowSOM}}, or
\code{\link[=clusterSubsetWithPCA]{clusterSubsetWithPCA()}}.}

\item{sample_df}{A data frame of sample info, as generated by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{design}{A design matrix as generated by \code{\link[=makeDesignMatrix]{makeDesignMatrix()}}.}

\item{contrasts}{A contrasts matrix as generated by \code{\link[=makeContrastsMatrix]{makeContrastsMatrix()}}.}

\item{counts}{A count matrix as generated by \code{\link[=makeCountMatrix]{makeCountMatrix()}}.}

\item{markers_of_interest}{A character vector of markers to test for differential
expression.}

\item{meta_names}{A character vector of metacluster names to test. Default is
all metaclusters, \code{levels(fsom$metaclustering)}.}

\item{prepr_transform}{A filepath to the .rds file for the logicle
transformation applied during preprocessing. Default is
\code{file.path("RDS", "logicle_transformation.rds")}.}

\item{controls_df}{Optional, a data frame as generated by \code{\link[=prepareControlInfo]{prepareControlInfo()}},
containing information about FMO or isotype controls to be used to calculate
delta MFIs.}

\item{ctrl_fsom_names}{Optional, a character string defining names of FlowSOM
.rds files previously generated from \code{\link[=clusterControls]{clusterControls()}}.}

\item{subsetted_meta}{If \code{fsom} is a reclustering, and \code{controls_df}
is not \code{NULL}, a string or list of strings defining the metacluster(s)
\code{fsom} is a reclustering of.}
}
\value{
A list of lists containing test results and data matrices.
}
\description{
Perform differential expression analysis on markers of interest.
}
\details{
Additional details...

!!! counts parameter may be useless, Group column in \code{df_full} no longer needed
}
