% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_agnostic.R
\name{editTableMetaclusters}
\alias{editTableMetaclusters}
\title{editTableMetaclusters}
\usage{
editTableMetaclusters(
  table,
  new_labels = NULL,
  cluster_assignments = NULL,
  level_order = NULL
)
}
\arguments{
\item{table}{A data table that has been clustered by \code{\link[=flowSOMWrapper]{flowSOMWrapper()}}.}

\item{new_labels}{Optional. A named vector, where names are the original metacluster names,
and values are the new labels.}

\item{cluster_assignments}{Optional. A named vector, where names are the cluster numbers,
and values are metacluster names.}

\item{level_order}{A character vector giving the desired order of the metacluster labels.}
}
\value{
A data.table with an added column \code{Meta_edited}.
}
\description{
Merge or relabel metaclusters, and reassign clusters.
}
