% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{filterData}
\alias{filterData}
\title{filterData}
\usage{
filterData(input, clusters = NULL, metaclusters = NULL)
}
\arguments{
\item{input}{A data table or FlowSOM object.}

\item{clusters}{A character vector specifying which clusters to keep. Default
is \code{NULL}.}

\item{metaclusters}{A character vector specifying which metaclusters to keep.
Default is \code{NULL}.}
}
\value{
A data.table or flowFrame, where only cells belonging to the given clusters or
metaclusters are included.
}
\description{
filterData
}
\keyword{internal}
