% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_agnostic.R
\name{flowSOMWrapper}
\alias{flowSOMWrapper}
\title{flowSOMWrapper}
\usage{
flowSOMWrapper(
  table,
  cols_to_cluster,
  num_clus,
  seed = NULL,
  fsom_file = NULL,
  ...
)
}
\arguments{
\item{table}{A data.table, as generated by either \code{\link[=getTableFromFCS]{getTableFromFCS()}} or
\code{\link[=doPreprocessing]{doPreprocessing()}}.}

\item{cols_to_cluster}{A numeric or character vector specifying which columns
should be used for clustering.}

\item{num_clus}{The number of metaclusters to create.}

\item{seed}{Optional, a seed for reproducibility. Default is \code{NULL}.}

\item{fsom_file}{An .rds filename. If not \code{NULL} (default), the resulting
FlowSOM object will be saved under this name.}

\item{...}{Additional parameters to pass to \code{\link[FlowSOM:FlowSOM]{FlowSOM()}}.}
}
\value{
A data.table annotated with cluster and metacluster assignments.
}
\description{
flowSOMWrapper
}
