% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{getExprMatDE}
\alias{getExprMatDE}
\title{getExprMatDE}
\usage{
getExprMatDE(fsom_dt, marker_cols)
}
\arguments{
\item{fsom_dt}{A data.table which columns for markers/channels, a column \code{File}
denoting the sample a cell is from, and a column \code{Metacluster} denoting the
metacluster a cell belongs to}

\item{marker_cols}{A character vector of markers/channels of interest; these
should be column names of \code{fsom_dt}}
}
\value{
A data.table where columns are samples and rows are metacluster/marker groups
}
\description{
Find expression matrix: metacluster.marker by sample
}
