% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getSampleMetaclusterMFIs}
\alias{getSampleMetaclusterMFIs}
\title{getSampleMetaclusterMFIs}
\usage{
getSampleMetaclusterMFIs(
  input,
  col,
  sample_df,
  populations = NULL,
  inverse = FALSE
)
}
\arguments{
\item{input}{description}

\item{col}{The channel to find sample-metacluster MFIs for.}

\item{sample_df}{If \code{input} is a FlowSOM object, a data frame from
\code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{populations}{A \code{character} list of the names of metaclusters/populations to calculate MFIs for.}

\item{inverse}{Boolean, whether data should be back-transformed before calculating MFIs.
Only valid when \code{input} is a \code{GatingSet} containing a transformation.}
}
\value{
A table, where rows are samples and columns are metaclusters.
}
\description{
Get a table of MFIs, where each row is a sample and each column is a metacluster.
}
