% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getSampleMetaclusterMFIs}
\alias{getSampleMetaclusterMFIs}
\title{getSampleMetaclusterMFIs}
\usage{
getSampleMetaclusterMFIs(input, col, csv_name = NULL)
}
\arguments{
\item{input}{description}

\item{col}{The channel to find sample-metacluster MFIs for.}

\item{csv_name}{Optional. If you would like the resulting table to be saved
to a .csv file, define the desired filename here.}
}
\value{
A table, where rows are samples and columns are metaclusters.
}
\description{
Get a table of MFIs, where each row is a sample and each column is a metacluster.
}
