% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R, R/utility_functions.R
\name{getSampleMetaclusterMFIs}
\alias{getSampleMetaclusterMFIs}
\title{getSampleMetaclusterMFIsOld}
\usage{
getSampleMetaclusterMFIs(input, col, csv_name = NULL)

getSampleMetaclusterMFIs(input, col, csv_name = NULL)
}
\arguments{
\item{input}{description}

\item{col}{The channel to find sample-metacluster MFIs for.}

\item{csv_name}{Optional. If you would like the resulting table to be saved
to a .csv file, define the desired filename here.}

\item{df_full}{A data frame with sample/metacluster information for the current
marker.}

\item{col_to_use}{Which column of the data frame should be used to create the
matrix.}
}
\value{
A matrix, where each column is a metacluster and each row is a sample.
The entries may be median expression on either a linear or logicle-transformed
scale.

A table, where rows are samples and columns are metaclusters.
}
\description{
Helper function for creating matrices of interest from \code{df_full}.

Get a table of MFIs, where each row is a sample and each column is a metacluster.
}
\keyword{internal}
