% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeCountMatrix}
\alias{makeCountMatrix}
\title{makeCountMatrix}
\usage{
makeCountMatrix(fsom, sample_df, meta_names)
}
\arguments{
\item{fsom}{A FlowSOM object.}

\item{sample_df}{A data frame containing sample information, generated either
manually or by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{meta_names}{A vector of metacluster names of interest.}
}
\value{
A matrix, where each column represents a sample, and each row
represents a metacluster.
}
\description{
Generate matrix of sample/metacluster cell counts.
}
