% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeCountMatrix}
\alias{makeCountMatrix}
\title{makeCountMatrix}
\usage{
makeCountMatrix(
  fsom,
  factors,
  meta_names = levels(fsom$metaclustering),
  min_cells = 3,
  min_samples = NULL
)
}
\arguments{
\item{fsom}{A FlowSOM object.}

\item{factors}{A data frame containing sample information, generated by
\code{\link[=makeFactorDF]{makeFactorDF()}}.}

\item{meta_names}{A vector of metacluster names of interest. By default, all
metacluster names are used.}

\item{min_cells}{An integer, the minimum number of cells a metacluster should
have in a specified number of samples to be included in the analysis.}

\item{min_samples}{An integer, the minimum number of samples a metacluster
should have at least \code{min_cells} events in to be included in the analysis.
By default, this is half the total number of samples.}
}
\value{
A matrix, where each column represents a sample, and each row
represents a metacluster.
}
\description{
Generate matrix of sample/metacluster cell counts.
}
