% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeDesignMatrix}
\alias{makeDesignMatrix}
\title{makeDesignMatrix}
\usage{
makeDesignMatrix(factors, comparisons)
}
\arguments{
\item{factors}{A data frame, generated either manually or by \code{\link[=makeFactorDF]{makeFactorDF()}}.}

\item{comparisons}{A named list of named lists, defining the groups to be
compared during analysis. See example for how this variable should be defined.}
}
\value{
A matrix where each column is a group of interest, and each row is a
sample.
}
\description{
Generate a design matrix. (edit example)
}
\examples{
samples <- prepareSampleInfo("filepath", "Sample.Name", "File.Name", c("Sex", "Disease"))

comparisons <- list(
  male_vs_female = list(Sex = list("male", "female")),
  male_vs_female_mibc = list(Disease = "MIBC", Sex = list("male", "female"))
  )

factors <- makeFactorDF(sample_df, comparisons)

design <- makeDesignMatrix(factors, comparisons)
}
