% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{makeFactorDF}
\alias{makeFactorDF}
\title{makeFactorDF}
\usage{
makeFactorDF(sample_df, comparisons)
}
\arguments{
\item{sample_df}{A data frame containing sample information, generated either
manually or by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{comparisons}{A named list of named lists, defining the groups to be
compared during analysis. See example for how this variable should be defined.}
}
\value{
A data frame of factors, with an added column for group.
}
\description{
! consider merging with prepareSampleInfo()
}
