% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{overwriteMetaclusterNames}
\alias{overwriteMetaclusterNames}
\title{overwriteMetaclusterNames}
\usage{
overwriteMetaclusterNames(fsom_dt, fsom_sub)
}
\arguments{
\item{fsom_dt}{Table with all cells of interest}

\item{fsom_sub}{Subset of table used for reclustering

Both tables should have a \code{cell_id} and \code{Metacluster} column.}
}
\value{
A table with updated metacluster names from the reclustering.
}
\description{
overwriteMetaclusterNames
}
