% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{plotGroupMFIBars}
\alias{plotGroupMFIBars}
\title{plotGroupMFIBars}
\usage{
plotGroupMFIBars(
  input,
  sample_df,
  comparison,
  meta_to_plot = NULL,
  upper_lim = NULL
)
}
\arguments{
\item{input}{A matrix, data frame, or path to a .csv file, where each row is
a sample and each column is a metacluster.}

\item{sample_df}{A factor data frame as generated by \code{\link[=prepareSampleInfo]{prepareSampleInfo()}}.}

\item{comparison}{A list defining a comparison.}

\item{meta_to_plot}{A vector of strings defining the metaclusters to include
in the plot. Default is all metaclusters.}

\item{upper_lim}{The y-axis upper limit.}
}
\value{
A bar plot drawn with \code{\link[ggplot2]{ggplot2}}.
}
\description{
Function to draw bar plot of MFIs or dMFIs by group.
}
\details{
The names and order of the groups defined in \code{comparison} should
appear in the same way you would like them to appear in the bar plot. Make
sure that the group names you supply are the same as they appear in the
\code{sample_df} object.

!!! should be able to better use tidyr and dplyr here
}
