% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{plotGroupMFIBars}
\alias{plotGroupMFIBars}
\title{plotGroupMFIBars}
\usage{
plotGroupMFIBars(csv_file, factors, grps_of_interest, upper_lim = NULL)
}
\arguments{
\item{csv_file}{A .csv file.}

\item{factors}{A factor data frame as generated by \code{\link[=makeFactorDF]{makeFactorDF()}}.}

\item{grps_of_interest}{A list defining which groups to plot.}

\item{upper_lim}{The y-axis upper limit.}
}
\value{
A bar plot drawn with \code{\link[ggplot2]{ggplot2}}.
}
\description{
...
}
