% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotLabeled2DScatter}
\alias{plotLabeled2DScatter}
\title{plotLabeled2DScatter}
\usage{
plotLabeled2DScatter(
  input,
  channelpair,
  clusters = NULL,
  metaclusters = NULL,
  labels = TRUE,
  colors = NULL
)
}
\arguments{
\item{input}{A FlowSOM object or data table.}

\item{channelpair}{A vector of two channel names.}

\item{clusters}{A vector specifying clusters of interest.}

\item{metaclusters}{A vector specifying metaclusters of interest.}

\item{labels}{Logical, should labels be plotted for each cluster center. Default
is \code{TRUE}.}

\item{colors}{Optional, a vector of colors for each metacluster.}
}
\value{
A plot drawn with ggplot2.
}
\description{
Plot a 2D scatterplot colored by metacluster.
}
\details{
It is only necessary to define one of the arguments \code{clusters} and
\code{metaclusters}, but the user may use both if they wish to plot a particular
subset of their data.

This function draws a 2D scatterplot with the given channels on each axis.
Cells are colored by metacluster, and if \code{labels = TRUE}, a label is
plotted at each cluster center. The number contained in these labels indicates
the cluster's number, and the color of the label indicates which metacluster it
belongs to.
}
\examples{
# Read in data
file <- system.file("extdata", "fsom_table_init.rds", package = "flowFun")
fsom_dt <- readRDS(file)

# Generate plot
plotLabeled2DScatter(fsom_dt,
                     channelpair = c("APC-Cy7-A", "BUV615-P-A"),
                     metaclusters = c(17, 19))

# Generate same plot with no labels
plotLabeled2DScatter(fsom_dt,
                     channelpair = c("APC-Cy7-A", "BUV615-P-A"),
                     metaclusters = c(17, 19),
                     labels = FALSE)

# Specify both metaclusters and clusters
plotLabeled2DScatter(fsom_dt,
                     channelpair = c("APC-Cy7-A", "BUV563-A"),
                     clusters = c(66, 79, 91),
                     metaclusters = 16)

# Specifying only clusters still colors by metacluster
plotLabeled2DScatter(fsom_dt,
                     channelpair = c("APC-Cy7-A", "BUV563-A"),
                     clusters = c(66, 79, 91))
}
