% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotLabeled2DScatter}
\alias{plotLabeled2DScatter}
\title{plotLabeled2DScatter}
\usage{
plotLabeled2DScatter(
  fsom,
  channelpair,
  clusters = NULL,
  metaclusters = NULL,
  label_type = c("cluster", "metacluster")
)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by the FlowSOM function.}

\item{channelpair}{A vector specifying which two channels to plot.}

\item{clusters}{A vector of cluster indices.}

\item{metaclusters}{A vector of metacluster indices.}

\item{label_type}{A string specifying whether you would like the cluster
centers to be labeled with the cluster or metacluster they belong to, either
"cluster" or "metacluster".}
}
\value{
A scatterplot made with ggplot and \code{FlowSOM::Plot2DScatters}.
}
\description{
Plot 2D scatterplot of FlowSOM clusters, with (meta)cluster number labels for
each cluster center.
}
\examples{
file <- system.file("extdata", "sample_aggregate.fcs", package = "flowFun")
fsom <- FlowSOM::FlowSOM(file,
                         colsToUse = c(10, 12:14, 16, 18:23, 25:32, 34),
                         nClus = 15,
                         seed = 42)

plotLabeled2DScatter(fsom,
                     c("APC-Cy7-A", "BUV563-A"),
                     metaclusters = c(3, 7, 10),
                     label_type = "cluster")
}
