% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_agnostic.R
\name{plotMetaclusterMFIsNew}
\alias{plotMetaclusterMFIsNew}
\title{Title}
\usage{
plotMetaclusterMFIsNew(input, cols_to_use = NULL, ...)
}
\arguments{
\item{input}{...}

\item{cols_to_use}{...}

\item{...}{}
}
\value{
...
}
\description{
Title
}
