% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotSampleProportions}
\alias{plotSampleProportions}
\title{plotSampleProportions}
\usage{
plotSampleProportions(fsom, reg_expr = NULL)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by \code{\link[FlowSOM:FlowSOM]{FlowSOM()}},
or \code{\link[=clusterSubsetWithPCA]{clusterSubsetWithPCA()}}.}

\item{reg_expr}{Optional, a regular expression to rename sample files.}
}
\value{
A stacked bar plot made with \code{\link[ggplot2]{ggplot2}}.
}
\description{
Plot cell type proportions by sample.
}
\details{
By default, the names used for each sample will be its file name. However,
you may rename the samples by providing a regular expression with a capturing
group, which will be applied to each file name.
}
