% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotUMAP}
\alias{plotUMAP}
\title{plotUMAP}
\usage{
plotUMAP(fsom, num_cells = 5000, labels = NULL, colors = NULL, seed = NULL)
}
\arguments{
\item{fsom}{A FlowSOM object as generated by the FlowSOM function.}

\item{num_cells}{The number of cells to use for the dimension reduction.
Default is 5000.}

\item{seed}{Optional, a seed for reproducibility.}
}
\value{
A UMAP plot drawn with ggplot2.
}
\description{
Plots UMAP colored by metacluster.
}
\examples{
file <- system.file("extdata", "sample_aggregate.fcs", package = "flowFun")
fsom <- FlowSOM::FlowSOM(file,
                         colsToUse = c(10, 12:14, 16, 18:23, 25:32, 34),
                         nClus = 10,
                         seed = 42,
                         xdim = 6,
                         ydim = 6)

plotUMAP(fsom)

plotUMAP(fsom)
}
