% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_agnostic.R
\name{plotUMAPNew}
\alias{plotUMAPNew}
\title{Title}
\usage{
plotUMAPNew(input, num_cells = 5000, seed = NULL)
}
\arguments{
\item{input}{...}

\item{num_cells}{...}

\item{seed}{...}
}
\value{
...
}
\description{
*parameter tweaking
}
