% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{prepareControls}
\alias{prepareControls}
\title{prepareControls}
\usage{
prepareControls(
  sample_df,
  ctrl_col,
  dir_clustr_ctrl,
  marker,
  orig_fsom,
  fsom_name = NULL,
  agg_name = NULL,
  subsetted_meta = NULL
)
}
\arguments{
\item{ctrl_col}{The name of the column containing control filenames.}

\item{dir_clustr_ctrl}{The directory containing preprocessed control files.}

\item{marker}{The marker that this control corresponds to.}

\item{orig_fsom}{The original FlowSOM object, as generated by \code{\link[FlowSOM]{FlowSOM}}.}

\item{fsom_name}{}
}
\value{
A FlowSOM object.
}
\description{
Prepare isotype or FMO controls for calculating delta MFI.
}
\details{
Additional details...
}
