% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_functions.R
\name{prepareSampleInfo}
\alias{prepareSampleInfo}
\title{prepareSampleInfo}
\usage{
prepareSampleInfo(
  filepath,
  name_col,
  filename_col,
  cols_to_use,
  samples_to_remove = NULL
)
}
\arguments{
\item{filepath}{A filepath to a .csv file containing sample info.}

\item{name_col}{The name of the column containing the experiment's sample names.}

\item{filename_col}{The name of the column containing the experiment's .fcs file names.}

\item{cols_to_use}{A character vector specifying the column names other than
sample name and filename that are relevant for the experiment.}

\item{samples_to_remove}{An integer vector defining the rows of any samples
that should be excluded from the analysis, either because they were discarded
during preprocessing or are not of interest. Default is none (\code{NULL}).}
}
\value{
A data frame containing sample information.
}
\description{
Prepare sample information for DE and DA analysis via a .csv file.
}
\details{
If you have a .csv file specifying sample information, it can be read in and
prepared by this function. It MUST have columns for file names and sample
names, and any additional columns may specify parameters of interest,
like NAC vs. No NAC. Each column should be named after what it represents.
This file will be read in as a data frame and will be used to construct the
design matrix used for analysis.

\emph{REMOVE} differences from orig. script: columns not of interest are removed,
rather than keeping all columns and storing that info as a variable. filename
remains in the data frame. all columns other than filename and sample names
have make.names() applied. instead of ordering and removing rows with
prepr_files, they are ordered alphabetically according to filename; this is
the same as how list.files() reads in filenames. \emph{add option to override?}
}
