% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_functions.R
\name{revertPCAFlowSOM}
\alias{revertPCAFlowSOM}
\title{revertPCAFlowSOM}
\usage{
revertPCAFlowSOM(input, fsom, pca_obj, num_components)
}
\arguments{
\item{input}{The aggregate flowFrame or data table used to generate the given FlowSOM object.}

\item{fsom}{A FlowSOM object as generated by \code{\link[=clusterSubsetWithPCA]{clusterSubsetWithPCA()}}.}

\item{pca_obj}{A \code{\link[stats]{prcomp}} object as generated by \code{\link[=doPCA]{doPCA()}}}

\item{num_components}{The number of principal components to use for reclustering,
as determined by examining results of \code{\link[=plotPCAScree]{plotPCAScree()}}.}
}
\value{
A FlowSOM object defined in terms of channels.
}
\description{
Reverts elements of a FlowSOM object generated on principal components back
to channels.
}
\details{
This function converts a FlowSOM object created from principal components,
as generated by \code{clusterSubsetWithPCA()}, into an
equivalent FlowSOM object defined in terms of channels instead. The
cluster/metacluster assignments for each cell remains the same, but the
resulting FlowSOM object is able to be used with built-in FlowSOM functions
like \code{PlotStars}, \code{PlotPies}, etc. Therefore other flowFun functions
like \code{plotLabeled2DScatter} and \code{searchByExpression} also work on
the resulting object.
}
\keyword{internal}
