% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tableToFlowSet}
\alias{tableToFlowSet}
\title{tableToFlowSet}
\usage{
tableToFlowSet(table, id_col = .id)
}
\arguments{
\item{table}{}
}
\value{
A flowSet
}
\description{
Convert a table of single-cell data
!!! from here, add function to save .fcs files
}
